+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 master
 ------
 
+* Reports can be rendered with jinja2 templates (to HTML, CSV or similar)
 * Manual file-based accounts can now also contain balances for any asset
 * New table output with better number formatting
 * Bugfix: Fiat currency setting was not correctly used in reports